feat: Support login with phone/email & OTP, require phone
This commit updates the user schema to make the `phone` field
required and the `email` field optional. It refactors the
authentication controller to support OTP verification and login via
email or phone number, including associated DTOs and service calls.

The AuthService is refactored to remove unnecessary comments and
adds a findByEmailOrPhone method to the usersService.  The OtpService
is updated to allow OTP generation and verification by either email
or phone, and adds a check to ensure users sign up and verify before
logging in with OTP.

The `CreateUserDto` is updated to make `email` optional and `phone`
required. The `createUser` and `findByEmail` methods in `UsersService`
are updated to support user identification by either `email` or
`phone`, preventing duplicate accounts based on either field. Also otp
generated for phone number with email. The changes also include the
token experation time .