feat: Implement auth & refactor conversations to use numbers
Implements authentication and authorization for the message module by
adding `@Auth()` and `@ApiBearerAuth()` decorators to the `ChatController`
and defining an `IUser` interface. It also modifies the service and
gateway to use the authenticated user's ID from the request and adds token
logging to the gateway.

Updates the conversation retrieval logic to find conversations regardless
of user order in `user1Id` and `user2Id` and refactors the getConversations
and getMessages to use number instead of string and also adds IUser
interface.